List available subobjects data

Utility function to list the available saved subobjects. This is the internal function which returns a version of the table with the full pathfiles and is not prettied or subsetted by type. See listSubObjectMini for the external version which provides those functions.

Usage

list_subobject_mini()